Move to python package for benching script
