+++ rustfst-python-bench/rustfst_python_bench/algorithms/connect.py
@@ -1,0 +1,12 @@
+class ConnectAlgorithm:
+
+    def __init__(self):
+        pass
+
+    @classmethod
+    def openfst_cli(cls):
+        return "fstconnect"
+
+    @classmethod
+    def rustfst_subcommand(cls):
+        return "connect"
