+++ rustfst-python-bench/rustfst_python_bench/algorithms/invert.py
@@ -1,0 +1,13 @@
+class InvertAlgorithm:
+
+    def __init__(self):
+        pass
+
+    @classmethod
+    def openfst_cli(cls):
+        return "fstinvert"
+
+    @classmethod
+    def rustfst_subcommand(cls):
+        return "invert"
+
