+++ rustfst-python-bench/rustfst_python_bench/algorithms/project.py
@@ -1,0 +1,11 @@
+class ProjectAlgorithm:
+    def __init__(self):
+        pass
+
+    @classmethod
+    def openfst_cli(cls):
+        return "fstproject"
+
+    @classmethod
+    def rustfst_subcommand(cls):
+        return "project"
