+++ rustfst-python-bench/rustfst_python_bench/constants.py
@@ -1,0 +1,2 @@
+OPENFST_BINS = './openfst-1.7.2/bin/'
+RUSTFST_CLI = './target/release/rustfst-cli'
